Add rotation label in level editor
When rotating a map object, a label with the current rotation now appears. Added it because you may want to have two map objects with the same rotation, so now you wouldn't have to eyeball it.

Also, gdx-freetype was added to the Desktop Gradle because Hiero wouldn't work without it.