Add support for the finish point in the level editor
Now if you play test a level made in the level editor, the finish point actually works. When the player enters it, the color scheme transitions to the dark one. If the player stays long enough in the finish point, the level will restart, as to indicate that the level was finished. I restart the level because the tester may want to test the level continuously.