Fix bug regarding the color scheme
In the level editor, if you enter the finish point, the color scheme will change. If you stay long enough in the finish point, the level will restart, and the color scheme will eventually reset to normal.

However, if you pressed the back button during the transition between the color schemes, then the color scheme would remain changed in the entire game (in the level editor, the level selector, the main menu, everywhere), because it was not reset.