Fix another color scheme bug
In the level editor, if you play tested the level, went back to the level editor, and play tested it again, the color scheme wouldn't reset when the player left the finish point. The problem was that when you left the play test section, the color scheme would reset. When it reset, however, it would copy the reference of the color scheme's colors, instead of copying their RGBA values.