Fix camera reset bug
When you left the play test section of the level editor, the camera wouldn't reset to the initial position in the editor. The problem had to do with the way the camera was made to initially follow the player in 99993b863c76b73c158bf7f9cb8c799e8539d660.