Fix bug regarding the undo/redo system
In the level editor, if you tapped on a button (e.g. for resizing, movement) it would register it as an action (even if nothing changed about the map object). If you pressed undo, it wouldn't undo the change made and go to the previous state (not that there was any change to undo), but it would, for example, move the object to a seemingly random position.

The problem was that undo/redo uses the initial properties and compares them to the new properties. However, the initial properties were set in a function dragStart(), which was called only when a drag gesture started. A tap is not considered a drag gesture, so the initial properties were never initialized, which resulted them to be 0.

Now setting of the initial values is done in touchDown(), which is called regardless.