Fix camera bug
A refactoring caused the levelEntity to be reset at the wrong time, which in turn caused its width and height in tiles to become 0. The system that makes the camera follow the player keeps the camera within the bounds of the map, and to do that it uses the previously reset map width and height, which caused the camera to always be at (0,0).