Fix weird bug
Sometimes, after loading a level, bullets would be inside platforms (you could see a bullet texture), and if you shot, you could eventually shoot with the platform (so instead of shooting with bullets, the bullet was a platform form the map). I guess that this could be an interesting mechanic?