Add case to leaving the editor
Now, if you want to leave the editor, are editing a new level, but it is the default one (i.e. you didn't modify it in any way, did not move anything, did not add anything), then you won't be asked if you want to save it (and it won't be automatically saved).