Somewhat fix moving platforms
A platform that moved diagonally caused the player to be pushed when the direction of movement changed.

Now it still does this, but the effect is less noticeable, as I linearly interpolate between the current velocity and the target velocity, avoiding a sudden change.

I am not sure if I can do better than this (to make the player not move at all).