Disable level editor, cycle through levels
Now the level editor is disabled on mobile, as it isn't in a good enough state. Will add it back when I polish it (and maybe add a way to share levels).
Also now the player can cycle through the levels, being able to directly go from the last level to the first level, and back (closes #53).