Add debug IAP on Desktop
Now if you buy something on Desktop, SHOW_ADS will be set to false (previously it remained true).

The reason is for testing how the UI will change after removing ads (i.e. the no-ads button should become full, just like when you rate the game, and the text in the pop-up should change to indicate that a payment was already made), even though ads can't be shown on Desktop.