Remove maximum ads per session
The reason I added such cap is because I read that you shouldn't show more than 3-4 ads per session. So I just capped the maximum number of ads per session to 3, but what this really means (I think) is that in an average session (of an average length) there shouldn't be more than 3-4 ads.