Cache leaderboard
I am now caching the entire shots leaderboard every 6 hours (not play-time) to storage. The reason is that this will let players see their ranks even when offline, but also because, in time, players may finish the levels in fewer moves than the previous minimum, which means that I'll have to update every other player's rank accordingly, increasing it by 1.

So this seems to be a sensible solution to this problem, instead of downloading the entire leaderboard every time the game is launched,