Hide buttons on last level
Every button in the top row (skip level, no ads, leaderboard and restart) is now hidden on the last level.

The skip level one was hidden anyways, but I wanted to hide the leaderboard one as well, and if I didn't hid the other ones, it would have looked too asymmetrical.